agonising warning

+++ src/lib_stable/2d_parity_matrix/of_2d_parity_api.c
@@ -431,8 +431,8 @@ of_status_t	of_2d_parity_get_control_parameter  (of_2d_parity_cb_t*	ofcb,
 	switch (type) {
 	case OF_CTRL_GET_MAX_K:
 		if (value == NULL || length != sizeof(UINT32)) {
-			OF_PRINT_ERROR(("%s: OF_CTRL_GET_MAX_K ERROR: null value or bad length (got %d, expected %ld)\n",
-				__FUNCTION__, length, sizeof(UINT32)))
+			OF_PRINT_ERROR(("%s: OF_CTRL_GET_MAX_K ERROR: null value or bad length (got %d, expected %d)\n",
+				__FUNCTION__, length, (int)sizeof(UINT32)))
 			goto error;
 		}
 		*(UINT32*)value = ofcb->max_nb_source_symbols;
@@ -441,8 +441,8 @@ of_status_t	of_2d_parity_get_control_parameter  (of_2d_parity_cb_t*	ofcb,
 
 	case OF_CTRL_GET_MAX_N:
 		if (value == NULL || length != sizeof(UINT32)) {
-			OF_PRINT_ERROR(("%s: OF_CTRL_GET_MAX_N ERROR: null value or bad length (got %d, expected %ld)\n",
-				__FUNCTION__, length, sizeof(UINT32)))
+			OF_PRINT_ERROR(("%s: OF_CTRL_GET_MAX_N ERROR: null value or bad length (got %d, expected %d)\n",
+				__FUNCTION__, length, (int)sizeof(UINT32)))
 			goto error;
 		}
 		*(UINT32*)value = ofcb->max_nb_encoding_symbols;
