+++ src/lib_stable/reed-solomon_gf_2_8/of_reed-solomon_gf_2_8_api.c
@@ -497,8 +497,8 @@ of_status_t	of_rs_get_control_parameter    (of_rs_cb_t*	ofcb,
 	switch (type) {
 	case OF_CTRL_GET_MAX_K:
 		if (value == NULL || length != sizeof(UINT32)) {
-			OF_PRINT_ERROR(("%s: OF_CTRL_GET_MAX_K ERROR: null value or bad length (got %d, expected %ld)\n",
-				__FUNCTION__, length, sizeof(UINT32)))
+			OF_PRINT_ERROR(("%s: OF_CTRL_GET_MAX_K ERROR: null value or bad length (got %d, expected %d)\n",
+				__FUNCTION__, length, (int)sizeof(UINT32)))
 			goto error;
 		}
 		*(UINT32*)value = ofcb->max_nb_source_symbols;
@@ -507,8 +507,8 @@ of_status_t	of_rs_get_control_parameter    (of_rs_cb_t*	ofcb,
 
 	case OF_CTRL_GET_MAX_N:
 		if (value == NULL || length != sizeof(UINT32)) {
-			OF_PRINT_ERROR(("%s: OF_CTRL_GET_MAX_N ERROR: null value or bad length (got %d, expected %ld)\n",
-				__FUNCTION__, length, sizeof(UINT32)))
+			OF_PRINT_ERROR(("%s: OF_CTRL_GET_MAX_N ERROR: null value or bad length (got %d, expected %d)\n",
+				__FUNCTION__, length, (int)sizeof(UINT32)))
 			goto error;
 		}
 		*(UINT32*)value = ofcb->max_nb_encoding_symbols;
